+++ src/nowiki/comment.ts
@@ -46,10 +46,7 @@ export abstract class CommentToken extends NowikiBaseToken {
 			return [];
 		}
 		const e = generateForSelf(this, {start}, 'unclosed-comment', Parser.msg('unclosed $1', 'HTML comment'));
-		e.fix = {
-			range: [e.endIndex, e.endIndex],
-			text: '-->',
-		};
+		e.fix = {range: [e.endIndex, e.endIndex], text: '-->'};
 		return [e];
 	}