+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - [`ImageParameterToken.size`](https://github.com/bhsd-harry/wikiparser-node/wiki/ImageParameterToken#size) when the parameter ends with `pxpx`
 - [`ImageParameterToken.link`](https://github.com/bhsd-harry/wikiparser-node/wiki/ImageParameterToken#link) when containing [`QuoteToken`](https://github.com/bhsd-harry/wikiparser-node/wiki/QuoteToken)
 - Image parameter parsing for `link=` followed by an invalid external link
+- Framed images and manual thumbnails are always unscaled
 
 **Changed**